[Refactor] Navigation bar sub components
