+++ addon/components/slds-navigation-bar/menu.js
@@ -1,5 +1,5 @@
 import Ember from 'ember';
-import layout from '../templates/components/slds-navigation-bar-menu';
+import layout from '../../templates/components/slds-navigation-bar/menu';
 
 export default Ember.Component.extend({
   layout,
