+++ addon/templates/components/slds-navigation-bar.hbs
@@ -15,8 +15,8 @@
   <nav class="slds-context-bar__secondary" role="navigation">
     <ul class="slds-grid">
       {{yield (hash 
-        item=(component 'slds-navigation-bar-item')
-        menu=(component 'slds-navigation-bar-menu')
+        item=(component 'slds-navigation-bar/item')
+        menu=(component 'slds-navigation-bar/menu')
       )}}
     </ul>
   </nav>
