+++ addon/templates/components/slds-navigation-bar/item.hbs
