+++ addon/templates/components/slds-navigation-bar/menu.hbs
@@ -15,5 +15,4 @@
   {{#slds-dropdown clicked=(action 'clicked') as |dropdown|}}
     {{yield dropdown}}
   {{/slds-dropdown}}
-
 </li>
