--- app/components/slds-navigation-bar-menu.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-cli-lightning-design-system/components/slds-navigation-bar-menu';
