+++ app/components/slds-navigation-bar/menu.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-cli-lightning-design-system/components/slds-navigation-bar/menu';
