+++ tests/dummy/app/templates/examples/navigation-bar.hbs
@@ -1,5 +1,5 @@
 {{#slds-navigation-bar applicationName='My Application' as |bar|}}
-  {{bar.action text='Home' path='examples.navigation-bar'}}
-  {{bar.action text='Events'}}
-  {{bar.action text='Properties'}}
+  {{bar.item text='Home' path='examples.navigation-bar'}}
+  {{bar.item text='Events'}}
+  {{bar.item text='Properties'}}
 {{/slds-navigation-bar}}
