+++ tests/integration/components/slds-navigation-bar/menu-test.js
@@ -1,11 +1,11 @@
 import { moduleForComponent, test } from 'ember-qunit';
 import hbs from 'htmlbars-inline-precompile';
 
-moduleForComponent('slds-navigation-bar-menu', 'Integration | Component | slds navigation bar menu', {
+moduleForComponent('slds-navigation-bar/menu', 'Integration | Component | slds navigation bar/menu', {
   integration: true
 });
 
 test('it renders', function(assert) {
   assert.expect(0);
-  this.render(hbs`{{slds-navigation-bar-menu}}`);
+  this.render(hbs`{{slds-navigation-bar/menu}}`);
 });
